changed identity columns to Long

+++ src/main/db/01_currency_snapshot_setup.sql
@@ -5,14 +5,14 @@ use crypto;
 drop table if exists currency_snapshot cascade;
 
 create table currency_snapshot (
-	Id int not null auto_increment,
+	Id bigint not null auto_increment,
     Name varchar(50) not null,
     Symbol varchar(10) not null,
     Rank int not null,
-    MarketCap long not null,
+    MarketCap bigint not null,
     Price decimal(10,2) not null,
-    CirculatingSupply long,
-    Volume_24h long,
+    CirculatingSupply bigint,
+    Volume_24h bigint,
     PercentChange_1h decimal(10,2),
     PercentChange_24h decimal(10,2),
     PercentChange_7d decimal(10,2),
